+++ Sachssoft.Sasopuls.Views.Shared/Inspector/IInspectorSchemaProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace Sachssoft.Sasopuls.Inspector
-{
-    public interface IInspectorSchemaProvider
-    {
-        IInspectorSchema Schema { get; }
-    }
-}