+++ Sachssoft.Sasopuls.Views.Shared/Messaging/CommandPolicy.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 
 namespace Sachssoft.Sasopuls.Messaging
 {